Rename "license" to "licence" across the codebase
Standardized spelling from "license" to "licence" in directories, packages, class names, and resources for consistency. Added a validation to throw an exception if no licences are found for a user. Updated a Docker command in the testing script to remove redundancy.
