+++ license-connect-authenticator/src/main/resources/META-INF/services/org.keycloak.authentication.AuthenticatorFactory
@@ -1,1 +1,1 @@
-de.intension.authentication.authenticators.license.LicenseConnectAuthenticatorFactory
+de.intension.authentication.authenticators.licence.LicenseConnectAuthenticatorFactory
