+++ license-connect-authenticator/src/main/resources/META-INF/services/org.keycloak.services.resource.RealmResourceProviderFactory
@@ -1,1 +1,1 @@
-de.intension.authentication.authenticators.license.rest.LicenceResourceProviderFactory
+de.intension.authentication.authenticators.licence.rest.LicenceResourceProviderFactory
